requirements.txt

Flask>=3.0.0
yara-python>=4.3.0
werkzeug>=3.0.0
